requirements.txt
pydub>=0.25.1
Pillow>=9.0.0
numpy>=1.21.0 